working useQuery
