+++ src/lib/App.tsx
@@ -3,6 +3,7 @@ import Localist from "./localist";
 import EventsState from "./js/context/EventState";
 import { AppProps } from "./types/types";
 import { QueryClient, QueryClientProvider } from "react-query";
+import { ReactQueryDevtools } from "react-query/devtools";
 
 const queryClient = new QueryClient();
 
@@ -11,6 +12,7 @@ const App = (props: AppProps) => (
     <EventsState>
       <Localist {...props} />
     </EventsState>
+    <ReactQueryDevtools initialIsOpen={true} />
   </QueryClientProvider>
 );
